requirements.txt
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
numpy==2.2.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pytesseract==0.3.13
sniffio==1.3.1
starlette==0.46.0
typing_extensions==4.12.2
uvicorn==0.34.0
